% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{acf_exact}
\alias{acf_exact}
\title{Function calculating the exact autocorrelation function of the
INARMA(p, q) model with general innovation distribution and parameters
\eqn{\kappa}, \eqn{\beta}, \eqn{\tau} and \eqn{\sigma^2_\tau}}
\usage{
acf_exact(max_d, beta, kappa, tau = NULL, sigma2_tau = NULL)
}
\arguments{
\item{max_d}{maximum lag}

\item{beta, kappa}{parameters of the model, both `beta` and `kappa` can be
vectors. Each of them must sum up to less than 1.}
}
\value{
vector of the autocorrelation values from lag 0 to `max_d`
}
\description{
Function calculating the exact autocorrelation function of the
INARMA(p, q) model with general innovation distribution and parameters
\eqn{\kappa}, \eqn{\beta}, \eqn{\tau} and \eqn{\sigma^2_\tau}
}
