% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{acf_exact_11}
\alias{acf_exact_11}
\title{Function calculating the exact autocorrelation function of INARMA(1, 1) with
general innovations.}
\usage{
acf_exact_11(max_d, beta, kappa, tau, sigma2_tau)
}
\arguments{
\item{max_d}{maximum lag}

\item{beta, kappa, tau, sigma2_tau}{parameters of the model. For `sigma2_tau = tau`
we get the Poisson model.}
}
\value{
vector of the autocorrelation values from lag 0 to `max_d`
}
\description{
Function calculating the exact autocorrelation function of INARMA(1, 1) with
general innovations.
}
