% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{acf_exact_ingarch}
\alias{acf_exact_ingarch}
\title{Function calculating the exact autocorrelation fun of the Poisson
INGARCH(p, q) model.}
\usage{
acf_exact_ingarch(max_d, beta, kappa, tau)
}
\arguments{
\item{max_d}{maximum lag}

\item{beta, kappa, tau}{model parameters in the thinning formulation}
}
\value{
vector of the autocorrelation values from lag 0 to `max_d`
}
\description{
The function uses the functionality from the `tscount` package.
}
\details{
function arguments:
}
