% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{acf_exact_pois}
\alias{acf_exact_pois}
\title{Function calculating the exact autocorrelation function of the Poisson
INARMA(p, q) model with parameters \eqn{\kappa} and \eqn{\beta}}
\usage{
acf_exact_pois(max_d, beta, kappa)
}
\arguments{
\item{max_d}{maximum lag}

\item{beta, kappa}{parameters of the model, both `beta` and `kappa` can be
vectors. Each of them must sum up to less than 1.}
}
\value{
vector of the autocorrelation values from lag 0 to `max_d`
}
\description{
Function calculating the exact autocorrelation function of the Poisson
INARMA(p, q) model with parameters \eqn{\kappa} and \eqn{\beta}
}
