% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{acov_exact}
\alias{acov_exact}
\title{Function calculating the exact (analytical) autocovariance function of the
INARMA(p, q) model.}
\usage{
acov_exact(max_d, beta, kappa, tau = NULL, sigma2_tau = NULL)
}
\arguments{
\item{max_d}{maximum lag}

\item{beta, kappa, tau, sigma2_tau}{parameters of the model, both `beta` and `kappa` can be
vectors. Each of them must sum up to less than 1.}
}
\value{
list with 2 elements
\describe{
\item{acov}{vector of the autocovariance values from lag 0 to `max_d`.}
\item{approx_OK}{indicator, whether the error of the infinite sum evaluation falls below `1e-8`.
For the Poisson case it is always `TRUE`.}
}
}
\description{
The model parameters are \eqn{\kappa}, \eqn{\beta}, \eqn{\tau} and
\eqn{\sigma^2_\tau}. For the non-Poisson case (\eqn{\tau} != \eqn{\sigma^2_\tau}),
an infinite sum must be evaluated. The precision of this evaluation is typically
below `1e-8`.
}
