% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_approx.R
\name{ar_approx_11}
\alias{ar_approx_11}
\title{Function calculating the AR-approximation for an INARMA(1, 1) model.}
\usage{
ar_approx_11(X, lag_max, family, control_optim = NULL)
}
\arguments{
\item{X}{a vector of observed count values}

\item{lag_max}{the order of the approximate AR model.}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{control_optim}{a list of options passed to `optim`}
}
\value{
an object of class `inarma`. This is a list with the following elements:
\describe{
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{fitted_values}{fitted values calculated from the approximate AR model}
\item{fitted_variance}{the fitted conditional variances calculated from the
approximate AR model. In the case of the AR model, they are just the estimate
 of the white noise variance}
\item{pearson_residuals}{the Pearson residuals, again obtained from the AR model}
\item{dim}{the number of fitted parameters}
\item{loglikelihood}{the log-likelihood of the Approximate model}
\item{AIC}{the resulting (approximate) AIC}
\item{convergence}{indicates whether optimization converged}
\item{optim}{return object of the call to `optim`}
}
}
\description{
Function calculating the AR-approximation for an INARMA(1, 1) model.
}
