% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{choose_support}
\alias{choose_support}
\title{Choose the support for likelihood approximation depending on the observations and parameters}
\usage{
choose_support(observed, tau, phi, kappa, psi = NULL, family)
}
\arguments{
\item{observed}{the time series of observed values}

\item{tau, phi, kappa, psi}{the model parameters}

\item{family}{the distributional family (`"Poisson"`, `"Hermite"` or `"NegBin`)}
}
\value{
the support; a vector of consecutive integers starting at 0.
}
\description{
Choose the support for likelihood approximation depending on the observations and parameters
}
