% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_approximation_helpers.R
\name{delta_method_for_pairs}
\alias{delta_method_for_pairs}
\title{Delta method for calculating the standard errors of the 2D INARMA parameters
 on the original scale}
\usage{
delta_method_for_pairs(pair, cov_raw_part)
}
\arguments{
\item{pair}{a pair of transformed model parameters. Either
\eqn{g(\kappa_1, \kappa_2)}, or \eqn{g(\beta_1, \beta_2)}}

\item{cov_raw_part}{block of the covariance matrix corresponding to the transformed
parameter pair}
}
\value{
the standard errors of the parameter pair on the original scale
}
\description{
Delta method for calculating the standard errors of the 2D INARMA parameters
 on the original scale
}
