% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_hermite.R
\name{dherm}
\alias{dherm}
\title{Probability mass function of the Hermite distribution}
\usage{
dherm(x, a1 = NULL, a2 = NULL, mu = NULL, psi = NULL, log = FALSE)
}
\arguments{
\item{x}{the values at which to evaluate the probability mass function}

\item{a1, a2}{parameters in format a1, a2 (in this case only some recycling is done)}

\item{mu, psi}{parameters in format mu, psi}
}
\description{
Probability mass function of the Hermite distribution
}
