% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_moments.R
\name{fit_inar_moments}
\alias{fit_inar_moments}
\title{See `?fit_inarma_moments` for details on the model definition.}
\usage{
fit_inar_moments(observed, family = c("Poisson", "Hermite", "NegBin"))
}
\arguments{
\item{observed}{a vector of observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{parameterization}{the function internally works with a slightly different notation as}
}
\value{
A named list with the following elements.
\describe{
\item{coefficients}{the estimated coefficients.}
\item{coefficients_uncorrected}{the estimated model coefficients without any corrections
(i.e., coefficients can be negative).}
\item{nobs}{the number of observations.}
\item{fitting_method}{the method used to fit the model, here `"moments"`.}
}
}
\description{
The function implements some heuristics to ensure model parameters fall into the allowed ranges.
See Bracher and Sobolova (2024) for details.
}
\examples{
data("measles")
X <- measles$value
fit <- fit_inar_moments(X, family = "Poisson")
summary(fit)


}
