% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ingarch_type_models.R
\name{fit_inarch}
\alias{fit_inarch}
\title{Fitting an INARCH model}
\usage{
fit_inarch(
  observed,
  family = c("Poisson", "Hermite", "NegBin"),
  start = NULL,
  return_se = TRUE,
  control_optim = NULL
)
}
\arguments{
\item{observed}{a vector containing the observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{start}{initial values for the optimization routine (on the internal
scale; check the element `"optim"` of the return list).}

\item{return_se}{should standard errors be returned?}

\item{control_optim}{a list of options passed to `optim`}
}
\value{
a list with the following elements:
\describe{
\item{family}{the distribution family used.}
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{if `return_se == TRUE`: the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{observed}{the vector of observed values provided by the user.}
\item{fitted_values}{the fitted values as obtained from `lik_distr`}
\item{pearson_residuals}{the Pearson residuals}
\item{dim}{the number of fitted parameters}
\item{loglikelihood}{the log-likelihood of the fitted model}
\item{AIC}{the resulting AIC}
\item{convergence}{indicates whether optimization converged}
\item{nobs}{the number of observations}
\item{optim}{return object of the call to `optim`}
\item{fitting_method}{the method used to fit the model, here `"maximum_likelihood"`.}
}
}
\description{
Maximum likelihood inference for INARCH(1) model, which is obtained as a
special case of the INGARCH(1, 1) model as described in Bracher and Sobolova
(2024) for \eqn{\beta = 0}. For the definition of the INGARCH(1, 1) model,
see `?fit_ingarch`.
}
\examples{
data("measles")
X <- measles$value
\dontrun{
fit <- fit_inarch(X, family = "Poisson")
}


}
