% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_approx.R
\name{fit_inarma_approx}
\alias{fit_inarma_approx}
\title{Fitting an INARMA model using the approximation by an AR model}
\usage{
fit_inarma_approx(
  observed,
  family = c("Poisson", "Hermite", "NegBin"),
  offspring = "binomial",
  order = c(p = 1, q = 1),
  lag_max = 10,
  control_optim = NULL
)
}
\arguments{
\item{observed}{a vector of observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{offspring}{the offspring distribution, or in other words, the type of
the \eqn{\kappa \bullet X_t} thinning; the AR-approximation procedure
supports only `"binomial"`.}

\item{order}{a vector with two named elements `p` and `q` denoting the order
of the INARMA(p, q) model to be fitted. p,q <= 2}

\item{lag_max}{the order of the approximate AR model.}

\item{control_optim}{a list of options passed to `optim`}
}
\value{
an object of class `inarma`. This is a list with the following elements:
\describe{
\item{family}{the distribution family used.}
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{observed}{the vector of observed values provided by the user.}
\item{lik_distr}{`NULL` value for the AR-approximation}
\item{fitted_values}{fitted values calculated from the approximate AR model}
\item{fitted_variance}{the fitted conditional variances calculated from the
approximate AR model. In the case of the AR model, they are just the estimate
 of the white noise variance}
\item{pearson_residuals}{the Pearson residuals, again obtained from the AR model}
\item{dim}{the number of fitted parameters}
\item{loglikelihood}{the log-likelihood of the Approximate model}
\item{AIC}{the resulting (approximate) AIC}
\item{convergence}{indicates whether optimization converged}
\item{nobs}{the number of observations}
\item{optim}{return object of the call to `optim`}
\item{order}{order of the fitted INARMA model.}
\item{fitting_method}{the method used to fit the model, here `"AR-approximation"`.}
}
}
\description{
Inference for INARMA(p, q) model (p,q <= 2) using the Ar-approximation as
described in Bracher and Sobolova (2025).
}
\details{
For the model definition see the `?fit_inarma` page of the documentation.

The function uses a moment matching procedure to match the INARMA and AR
model of order `lag_max`.
}
\examples{
data("measles")
X <- measles$value
# Note: running the fit takes a little while.
\dontrun{
fit <- fit_inarma_approx(X, family = "Poisson", order = c(p = 1, q = 1))
summary(fit)
plot(fit, type = "fit")
}


}
