% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_moments.R
\name{fit_inarma_moments}
\alias{fit_inarma_moments}
\title{Fitting an INARMA(1,1) model via method of moments.}
\usage{
fit_inarma_moments(
  observed,
  family = c("Poisson", "Hermite", "NegBin"),
  parameterization = "beta"
)
}
\arguments{
\item{observed}{a vector of observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{parameterization}{the function internally works with a slightly different notation as}
}
\value{
A named list with the following elements.
\describe{
\item{coefficients}{the estimated coefficients.}
\item{coefficients_uncorrected}{the estimated model coefficients without any corrections
(i.e., coefficients can be negative).}
\item{nobs}{the number of observations.}
\item{fitting_method}{the method used to fit the model, here `"moments"`.}
}
}
\description{
See `?fit_inarma` for details on the model definition.
}
\details{
The function implements some heuristics to ensure model parameters fall into the allowed ranges.
See Bracher and Sobolova (2024) for details.
}
\examples{
data("measles")
X <- measles$value
# Note: running the fit takes a little while.
fit <- fit_inarma_moments(X, family = "Poisson")
summary(fit)


}
