% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.inarma}
\alias{fitted.inarma}
\title{Get fitted values of an INARMA(p, q) model}
\usage{
\method{fitted}{inarma}(model)
}
\arguments{
\item{model}{the model (an object of class `"inarma"`)}
}
\description{
Get fitted values of an INARMA(p, q) model
}
