% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ingarch_type_models.R
\name{get_cluster_size}
\alias{get_cluster_size}
\title{Compute the clustersize and its standard error from the  model parameter psi
and the distribution family}
\usage{
get_cluster_size(psi, psi_se, family = c("Poisson", "Hermite", "NegBin"))
}
\arguments{
\item{psi}{the estimate of the parameter \eqn{psi} of the compound Poisson
distribution in the INGARCH framework}

\item{psi_se}{the standard error of \eqn{psi}}

\item{family}{family the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}
}
\value{
a list containing the estimate of the parameter `theta` and its
standard error `theta_se`
}
\description{
Compute the clustersize and its standard error from the  model parameter psi
and the distribution family
}
