% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_approximation_helpers.R
\name{get_ses_orig}
\alias{get_ses_orig}
\title{Function calculating the standard errors of all INARMA parameters on the
original scale}
\usage{
get_ses_orig(
  pars,
  hessian_transformed,
  family = c("Poisson", "Hermite", "NegBin")
)
}
\arguments{
\item{pars}{transformed model parameters, \eqn{\tau} on the log-scale, \eqn{\kappa} and
\eqn{\beta} on the logit-scale if fitting INARMA(1, 1), or on the 2D plane if
fitting a 2nd order INARMA}

\item{hessian_transformed}{hessian matrix from the optimization corresponding
to the transformed parameters}
}
\value{
a list of quantities
\describe{
\item{ses}{standard errors of parameters \eqn{\tau}, \eqn{\kappa} and
 \eqn{\beta} on the original scale}
\item{cov_raw}{the covariance matrix of the parameter estimate on the transformed scale.}
}
}
\description{
Function calculating the standard errors of all INARMA parameters on the
original scale
}
