% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood_approx.R
\name{llik_ar_based_11}
\alias{llik_ar_based_11}
\title{Function for evaluating the likelihood of an AR model approximating the
INARMA(1, 1) model}
\usage{
llik_ar_based_11(
  pars,
  X,
  lag_max = 8,
  return_fitted = FALSE,
  family = c("Poisson", "Hermite", "NegBin")
)
}
\arguments{
\item{pars}{transformed model parameters, \eqn{\kappa} and \eqn{\beta} on the
logit-scale, tau on the log-scale}

\item{X}{the observed count series}

\item{lag_max}{order of the AR model to be used for the approximation}

\item{family}{family of the INARMA model; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}
}
\value{
the likelihood of the approximating AR model
}
\description{
Function for evaluating the likelihood of an AR model approximating the
INARMA(1, 1) model
}
