% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood_approx.R
\name{llik_ar_based_higher}
\alias{llik_ar_based_higher}
\title{Function for evaluating the likelihood of an AR model approximating the
INARMA(p, q) model, for p, q <= 2}
\usage{
llik_ar_based_higher(
  pars,
  X,
  lag_max = 10,
  return_fitted = FALSE,
  family = c("Poisson", "Hermite", "NegBin"),
  return_approx_OK = FALSE
)
}
\arguments{
\item{pars}{transformed model parameters, \eqn{\kappa} and \eqn{\beta} on the
unconstrained scales (2D plane), \eqn{\tau} on the log-scale}

\item{X}{the observed count series}

\item{lag_max}{order of the AR model to be used for the approximation}
}
\value{
the likelihood of the approximating AR model
}
\description{
Function for evaluating the likelihood of an AR model approximating the
INARMA(p, q) model, for p, q <= 2
}
