% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_hingarch}
\alias{llik_hingarch}
\title{Evaluate log-likelihood for a Hermite INGARCH(1, 1) model}
\usage{
llik_hingarch(vect, tau, kappa, beta, psi, mean_E1, return_fitted = FALSE)
}
\arguments{
\item{vect}{the vector of observed values}

\item{tau, kappa, beta, psi, mean_E1}{parameters of the Hermite INGARCH(1,1)
model in the thinning-based formulation}

\item{return_fitted:}{should fitted values be returned?}
}
\value{
log-likelihood, or (if return_fitted) a list containing the log-likelihood
}
\description{
Evaluate log-likelihood for a Hermite INGARCH(1, 1) model
}
