% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_inarma0_tv}
\alias{llik_inarma0_tv}
\title{Generic function for likelihood evaluation}
\usage{
llik_inarma0_tv(
  vect,
  distr_E1,
  distr_I,
  phi,
  kappa,
  zeta,
  offspring,
  log = TRUE,
  return_distr = FALSE
)
}
\arguments{
\item{vect}{the vector of observed values}

\item{distr_E1}{a vector of probabilities used to initialize E1}

\item{distr_I}{the immigration distribution (probabilities for 0, ..., M); also implies the support for E and X}

\item{phi, kappa, zeta}{the model parameters}

\item{offspring}{the offspring distribution, or in other words, the type of
the \eqn{\kappa \bullet X_t} thinning; one of `"binomial"`, or `"binomial-Poisson"`}

\item{log}{should log-likelihood rather than likelihood be returned?}

\item{return_distr}{should a matrix with conditional probabilities (given the past) for each time point be returned?}
}
\value{
the conditional (log) likelihood as a numeric or a matrix containing all relevant conditional probabilities (if return_distr == TRUE)
}
\description{
Generic function for likelihood evaluation
}
