% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_inarma_pois}
\alias{llik_inarma_pois}
\title{Evaluating the likelihood in the Poisson case
A wrapper around llik_inarma0_tv}
\usage{
llik_inarma_pois(
  vect,
  distr_E1 = NULL,
  tau,
  phi,
  kappa,
  zeta,
  offspring,
  support = NULL,
  return_distr = FALSE
)
}
\arguments{
\item{vect}{the vector of observed values}

\item{distr_E1}{a vector of probabilities used to initialize E1}

\item{tau}{model parameter, scalar or a vector for time-varying immigration distributions}

\item{phi, kappa, zeta}{the model parameters, scalar}

\item{offspring}{the offspring distribution, or in other words, the type of
the \eqn{\kappa \bullet X_t}}

\item{return_distr}{should a matrix with conditional probabilities (given the past) for each time point be returned?}

\item{log}{should log-likelihood rather than likelihood be returned?}
}
\value{
the conditional (log) likelihood as a numeric or a matrix containing all relevant conditional probabilities (if return_distr == TRUE)
}
\description{
Evaluating the likelihood in the Poisson case
A wrapper around llik_inarma0_tv
}
