% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_ingarch}
\alias{llik_ingarch}
\title{Evaluate log-likelihood of a Poisson INGARCH(1,1) model}
\usage{
llik_ingarch(vect, tau, kappa, beta, mean_E1, return_fitted = FALSE)
}
\arguments{
\item{vect}{the vector of observed values}

\item{tau, kappa, beta, mean_E1}{parameters of the Poisson INGARCH(1,1)
model in the thinning-based formulation}

\item{return_fitted:}{should fitted values be returned?}
}
\value{
log-likelihood, or (if return_fitted) a list containing the log-likelihood
}
\description{
Evaluate log-likelihood of a Poisson INGARCH(1,1) model
}
