% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_nbingarch}
\alias{llik_nbingarch}
\title{Evaluate log-likelihood for a negative binomial INGARCH(1, 1) model}
\usage{
llik_nbingarch(vect, tau, kappa, beta, psi, mean_E1, return_fitted = FALSE)
}
\arguments{
\item{vect}{the vector of observed values}

\item{tau, kappa, beta, psi, mean_E1}{parameters of the Negative binomial
INGARCH(1,1) model in the thinning-based formulation}

\item{return_fitted:}{should fitted values be returned?}
}
\value{
log-likelihood, or (if return_fitted) a list containing the log-likelihood
}
\description{
Evaluate log-likelihood for a negative binomial INGARCH(1, 1) model
}
