% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measles}
\alias{measles}
\title{Bavarian measles data}
\format{
A data frame with 312 rows and 3 columns:
\describe{
  \item{year}{the year of reporting.}
  \item{week}{the week of reporting.}
  \item{value}{the cout value.}
}
}
\source{
<https://survstat.rki.de/>
}
\usage{
measles
}
\description{
Weekly numbers of reported measles cases in the German state of Bavaria, 2014-2019.
}
\keyword{datasets}
