% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_approximation_helpers.R
\name{orig_par_to_unconstrained}
\alias{orig_par_to_unconstrained}
\title{Transform a 2D parameter pair onto the unconstrained parametric space.}
\usage{
orig_par_to_unconstrained(par)
}
\arguments{
\item{par}{2D parameter of an INARMA model, either \eqn{(\kappa_1, \kappa_2)},
or  \eqn{(\beta_1, \beta_2)}}
}
\value{
transformed parameter (vector with two elements)
}
\description{
For the 2D parameter must hold which `par[1] + par[2] < 1` and
`0 < par[1], par[2] < 1`. The unconstrained parameter space is the whole
real plane.
}
