% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.inarma}
\alias{plot.inarma}
\title{Plotting functions for INARMA(1, 1) models}
\usage{
\method{plot}{inarma}(
  model,
  type = c("fit", "residuals", "racf"),
  interval_level = 0.8,
  residual_type = "Pearson",
  ylim = NULL,
  legend = TRUE,
  lag.max = 10
)
}
\arguments{
\item{model}{an object of class `"inarma"` to plot.}

\item{type}{the type of plot. One out of `"fit", "residuals", "racf"`.}

\item{interval_level}{for `type == "fit"`: the level for uncertainty intervals.}

\item{residual_type}{type of residuals to be used; one of `"raw"` or `"Pearson"`.}

\item{ylim}{the y-limit; only used for `type == "racf"`.}

\item{legend}{should a legend be shown? Only used for `type == "fitted`.}

\item{lag.max}{maximum lag to display if `type == "racf"`.}
}
\description{
Currently three plot types are available: model fit (`type = "fit"`),
residuals (`type = "residuals"`) and residual autocorrelation (`type = "racf"`).
}
\details{
Plots are done in the base R plotting framework.
}
