% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_inarma_fitted}
\alias{plot_inarma_fitted}
\title{Plot fitted values of INARMA model}
\usage{
plot_inarma_fitted(model, interval_level = 0.8, ylim = NULL, legend = TRUE)
}
\arguments{
\item{model}{an object of class `"inarma"` to plot.}

\item{interval_level}{a level for the uncertainty intervals (shaded area) displayed.}

\item{ylim}{a y-limit.}

\item{legend}{should a legend be shown?}
}
\description{
Plot fitted values of INARMA model
}
