% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_inarma_racf}
\alias{plot_inarma_racf}
\title{Plot residual autocorrelation function of an INARMA(1, 1) model}
\usage{
plot_inarma_racf(model, type = "Pearson", lag.max = 10)
}
\arguments{
\item{model}{an object of class `"inarma"`}

\item{type}{the type of residuals to be used; one of `"raw"` or `"Pearson"`.}
}
\description{
Plot residual autocorrelation function of an INARMA(1, 1) model
}
