% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_inarma_residuals}
\alias{plot_inarma_residuals}
\title{Plot residuals of an INARMA(1, 1) model}
\usage{
plot_inarma_residuals(model, type = c("raw", "Pearson"), ylim = NULL)
}
\arguments{
\item{model}{an object of class `"inarma"`.}

\item{type}{the type of residuals to be plotted; one of `"raw"` or `"Pearson"`.}
}
\description{
Plot residuals of an INARMA(1, 1) model
}
