% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.inarma}
\alias{print.inarma}
\title{Print an INARMA(p, q) model fit}
\usage{
\method{print}{inarma}(model)
}
\arguments{
\item{model}{the model fit (an object of class `"inarma"`).}
}
\description{
Print an INARMA(p, q) model fit
}
