% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_hermite.R
\name{rherm}
\alias{rherm}
\title{Sampling from the Hermite distribution}
\usage{
rherm(n, a1 = NULL, a2 = NULL, mu = NULL, psi = NULL)
}
\arguments{
\item{n}{the number of samples to generate}

\item{a1, a2}{parameters in format a1, a2 (in this case only some recycling is done)}

\item{mu, psi}{parameters in format mu, psi}
}
\description{
Sampling from the Hermite distribution
}
