% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_inarma_negbin}
\alias{sim_inarma_negbin}
\title{Simulate from a negative binomial INARMA(1,1) model}
\usage{
sim_inarma_negbin(
  tau,
  psi,
  phi,
  kappa,
  zeta,
  E1 = NULL,
  distr_E1 = NULL,
  lgt = NULL
)
}
\arguments{
\item{tau, psi, phi, kappa, zeta}{the model parameters.}

\item{E1}{an initial value for \eqn{E_1}.}

\item{distr_E1}{an initial distribution for \eqn{E_1}.}

\item{lgt}{the length of the simulated time series.}
}
\value{
A named list with the following elements:
\describe{
\item{X}{The process \eqn{X_t}}
\item{E}{The hidden process \eqn{E_t}}
\item{I}{The innovation process \eqn{I_t}}
}
}
\description{
Simulation from a negative binomial INARMA(1,1) model. For initialization an (approximation)
of the stationary distribution of \eqn{E[1]}. Typically used within the wrapper `sim_inarma`.
}
