% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_loop}
\alias{sim_loop}
\title{Execute the loop of the INARMA(1,1) model simulation}
\usage{
sim_loop(phi, kappa, zeta, I, E1, lgt)
}
\arguments{
\item{phi, kappa, zeta}{the model parameters.}

\item{I}{the sequence of imports/innovations with length `lgt`}

\item{E1}{an initial value for \eqn{E_1}.}

\item{lgt}{the length of the simulated time series.}
}
\value{
A named list with the following elements:
\describe{
\item{X}{The process \eqn{X_t}}
\item{E}{The hidden process \eqn{E_t}}
}
}
\description{
Execute the loop of the INARMA(1,1) model simulation
}
