% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_loop_higher_order}
\alias{sim_loop_higher_order}
\title{Execute the loop of the INARMA(p,q) model simulation, for
at least one p, q > 1}
\usage{
sim_loop_higher_order(phi, kappa, zeta, I, E1, lgt)
}
\arguments{
\item{phi, kappa, zeta}{the model parameters.}

\item{I}{the sequence of imports/innovations with length `lgt`}

\item{E1}{an initial values for \eqn{E_1}, ...\eqn{E_q} .}

\item{lgt}{the length of the simulated time series.}
}
\value{
A named list with the following elements:
\describe{
\item{X}{The process \eqn{X_t}}
\item{E}{The hidden process \eqn{E_t}}
}
}
\description{
Execute the loop of the INARMA(p,q) model simulation, for
at least one p, q > 1
}
