% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_properties.R
\name{sop_inarma}
\alias{sop_inarma}
\title{Compute the stationary second-order properties of an INARMA(1, 1) model.}
\usage{
sop_inarma(
  tau,
  psi = NULL,
  phi = NULL,
  beta = NULL,
  kappa,
  family = c("Poisson", "Hermite", "NegBin")
)
}
\arguments{
\item{tau, psi, phi, kappa}{the model parameters}

\item{family}{family the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}
}
\value{
a named list containing the following elements:
\describe{
\item{mu}{the stationary mean.}
\item{sigma2}{the stationary variance}
\item{acf1}{the autocorrelation \eqn{\rho_X(1)} at lag 1.}
\item{mu}{the decay factor of the autocorrelation function, i.e., \eqn{\rho_X(2)/\rho_X(1)}.}
}
}
\description{
Compute the stationary second-order properties of an INARMA(1, 1) model based on the model parameters.
}
