% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.inarma}
\alias{summary.inarma}
\title{Summarize an INARMA(p, q) model fit}
\usage{
\method{summary}{inarma}(model)
}
\arguments{
\item{model}{the model fit (an object of class `"inarma"`).}
}
\description{
Summarize an INARMA(p, q) model fit
}
