% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{tm_A_to_E_bulk}
\alias{tm_A_to_E_bulk}
\title{Calculate all necessary probabilities for tm_A_to_E in advance}
\usage{
tm_A_to_E_bulk(vect, kappa, zeta, support_E, offspring)
}
\arguments{
\item{vect}{the vector of observed values}

\item{kappa}{the model parameter kappa (offspring mean)}

\item{zeta}{the model parameter zeta; (0 corresponds to the binomial thinning,
1 to the Poisson thinning), values in between to the mixed thinning operator}

\item{support_E}{the chosen support for E}

\item{offspring}{the offspring distribution, or in other words, the type of
the \eqn{\kappa \bullet X_t}}
}
\value{
a matrix containing transition probabilities from all possible
X[t - 1] (values of E in rows, of X in columns); to be transformed into
transition matrices by function tm_A_to_E()
}
\description{
Calculate all necessary probabilities for tm_A_to_E in advance
}
