% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_hermite.R
\name{to_a1a2_herm}
\alias{to_a1a2_herm}
\title{Re-parameterize the Hermite distribution to a1, a2}
\usage{
to_a1a2_herm(a1 = NULL, a2 = NULL, mu = NULL, psi = NULL)
}
\arguments{
\item{a1, a2}{parameters in format a1, a2 (in this case only some recycling is done)}

\item{mu, psi}{parameters in format mu, psi; see Bracher and Sobolova (2024) for details.}
}
\description{
`a1, a2` are the parameters of the Hermite distribution in a parameterization
where \eqn{X \sim \text{Herm}(a_1, a_2)} is equivalent to
\deqn{X = Y_1 + 2Y_2, Y_1 \sim \text{Pois}(a_1), Y_2 \sim \text{Pois}(a_2).}
}
