% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_approximation_helpers.R
\name{unconstrained_par_to_orig}
\alias{unconstrained_par_to_orig}
\title{Back-transform a 2D parameter pair on the unconstrained space back to the
original space.}
\usage{
unconstrained_par_to_orig(par)
}
\arguments{
\item{par}{transformed 2D parameter of an INARMA model, either \eqn{g(\kappa_1, \kappa_2)},
or  \eqn{g(\beta_1, \beta_2)}}
}
\value{
back-transformed parameter (vector with two elements)
}
\description{
For the 2D parameter must hold which `par[1] + par[2] < 1` and
`0 < par[1], par[2] < 1`. The unconstrained parameter space is the whole
real plane.
}
