% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{var_exact_11}
\alias{var_exact_11}
\title{Function calculating the exact variance of the INARMA(1, 1) model with
general innovations.}
\usage{
var_exact_11(beta, kappa, tau, sigma2_tau)
}
\arguments{
\item{beta, kappa, tau, sigma2_tau}{parameters of the model. For `sigma2_tau = tau`
 we get the Poisson model.
\describe{
\item{fitting_method}{the method used to fit the model, here `"maximum_likelihood"`.}
}}

\item{max_d}{maximum lag}
}
\value{
variance of the INARMA(1, 1) process
}
\description{
Function calculating the exact variance of the INARMA(1, 1) model with
general innovations.
}
