% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_moments.R
\name{var_exact_ingarch}
\alias{var_exact_ingarch}
\title{Function calculating the exact variance of the Poisson INGARCH(p, q) model.}
\usage{
var_exact_ingarch(beta, kappa, tau)
}
\arguments{
\item{beta, kappa, tau}{model parameters in the thinning formulation. Both
`beta` and `kappa` can be vectors. Each of them must sum up to less than 1.}
}
\value{
the variance of the INGARCH(p, q)
}
\description{
The function uses the functionality from the `tscount` package.
}
\details{
function arguments:
}
