% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{AIC.inarma}
\alias{AIC.inarma}
\title{Get AIC of an INARMA(1, 1) model}
\usage{
\method{AIC}{inarma}(model)
}
\arguments{
\item{model}{the model (an object of class `"inarma"`)}
}
\description{
Get AIC of an INARMA(1, 1) model
}
