% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{check_arguments_llik}
\alias{check_arguments_llik}
\title{Checking the arguments of llik_inarma_pois, llik_inarma_herm, llik_inarma_nb}
\usage{
check_arguments_llik(
  vect,
  distr_E1,
  tau,
  phi,
  psi,
  kappa,
  support,
  log,
  return_distr
)
}
\arguments{
\item{vect}{the vector of observed values}

\item{distr_E1}{a vector of probabilities used to initialize E1}

\item{tau}{model parameter, scalar or a vector for time-varying immigration distributions}

\item{phi, kappa}{the model parameters, scalar}

\item{support}{the support for E and X}

\item{log}{should log-likelihood rather than likelihood be returned?}

\item{return_distr}{should a matrix with conditional probabilities (given the past) for each time point be returned?}
}
\value{
the conditional (log) likelihood as a numeric or a matrix containing all relevant conditional probabilities (if return_distr == TRUE)
}
\description{
Checking the arguments of llik_inarma_pois, llik_inarma_herm, llik_inarma_nb
}
