% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_inar}
\alias{fit_inar}
\title{Fitting an INAR model}
\usage{
fit_inar(
  observed,
  family = c("Poisson", "Hermite", "NegBin"),
  offspring = c("binomial", "binomial-Poisson"),
  start = NULL,
  return_se = TRUE,
  control_optim = NULL
)
}
\arguments{
\item{observed}{a vector of observed count values.}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`.}

\item{start}{initial values for the optimization routine.}

\item{return_se}{should standard errors be returned?}

\item{control_optim}{a list of options passed to `optim`.}
}
\value{
an object of class `inarma`. This is a list with the following elements:
\describe{
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{if `return_se == TRUE`: the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{observed}{the vector of observed values provided by the user.}
\item{lik_distr}{matrix containing for each time point the conditional probabilities for all
values in the support (given the past). This contains notably all likelihood contributions and
can serve to compute fitted values.}
\item{family}{the distribution family used.}
\item{fitted_values}{the fitted values as obtained from `lik_distr`.}
\item{fitted_variance}{the fitted conditional variances as obtained from `lik_distr`.}
\item{pearson_residuals}{the Pearson residuals.}
\item{dim}{the number of fitted parameters.}
\item{loglikelihood}{the log-likelihood of the fitted model.}
\item{AIC}{the resulting AIC.}
\item{convergence}{indicates whether optimization converged.}
\item{nobs}{the number of observations.}
\item{optim}{return object of the call to `optim`.}
\item{fitting_method}{the method used to fit the model, here `"maximum_likelihood"`.}
}
}
\description{
Maximum likelihood inference for INAR(1) model. The model is defined as
\deqn{X_t = \kappa \circ X_{t - 1} + I_t}
where \eqn{\circ} denotes binomial thinning. The immigration process \eqn{I_t} consists of independently
and identically distributed random variables which can be Poisson, Hermite or negative binomial.
This distribution is charaterized by its mean, denoted by \eqn{\tau}, and potentially a dispersion parameter \eqn{\psi}.
}
\details{
The function uses a variation of the forward algorithm to fit the models. It is a simplified version of
`fit_inarma` with the additional parameter \eqn{phi} set to 1.
}
\examples{
data("measles")
X <- measles$value
# Note: running the fit takes a little while.
\dontrun{
fit <- fit_inar(X, family = "Poisson")
}

}
