% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_inarma}
\alias{fit_inarma}
\title{Fitting an INARMA model}
\usage{
fit_inarma(
  observed,
  family = "Poisson",
  offspring = "binomial",
  parameterization = "beta",
  start = NULL,
  return_se = TRUE,
  control_optim = NULL
)
}
\arguments{
\item{observed}{a vector of observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{offspring}{the offspring distribution, or in other words, the type of
the \eqn{\kappa \bullet X_t} thinning; one of `"binomial"`, or `"binomial-Poisson"`}

\item{parameterization}{the function internally works with a slightly different notation as
the paper by Bracher and Sobolova (2024). The default `parameterization = "beta"` will return results
formated as in the paper. `parameterization = "phi"` returns results as handled internally,
with `phi = 1 - beta`.}

\item{start}{initial values for the optimization routine (on the internal scale; check the element `"optim"` of the return list)}

\item{return_se}{should standard errors be returned?}

\item{control_optim}{a list of options passed to `optim`}
}
\value{
an object of class `inarma`. This is a list with the following elements:
\describe{
\item{family}{the distribution family used.}
\item{offspring}{the offspring disribution used.}
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{if `return_se == TRUE`: the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{observed}{the vector of observed values provided by the user.}
\item{lik_distr}{matrix containing for each time point the conditional probabilities for all
values in the support (given the past). This contains notably all likelihood contributions and
can serve to compute fitted values.}
\item{fitted_values}{the fitted values as obtained from `lik_distr`}
\item{fitted_variance}{the fitted conditional variances as obtained from `lik_distr`}
\item{pearson_residuals}{the Pearson residuals}
\item{dim}{the number of fitted parameters}
\item{loglikelihood}{the log-likelihood of the fitted model}
\item{AIC}{the resulting AIC}
\item{convergence}{indicates whether optimization converged}
\item{nobs}{the number of observations}
\item{optim}{return object of the call to `optim`}
\item{fitting_method}{the method used to fit the model, here `"maximum_likelihood"`.}
}
}
\description{
Maximum likelihood inference for INARMA(1,1) model as described in Bracher and Sobolova (2025).
The model is defined as
\deqn{X_t = (1 - \beta) \circ E_t + I_t}
\deqn{E_t = \beta \circ E_{t - 1} + \kappa \bullet X_{t - 1}}
where \eqn{\circ} denotes binomial thinning and \eqn{\bullet} denotes either the
binomial thinning, or the binomial-Poisson thinning. The binomial-Poisson
thinning is defined as
\deqn{\kappa \bullet X_t = \sum_{k = 1}^{X_t} Z_k, \quad \text{where} \quad Z_k \sim \text{Pois}(\kappa\zeta) + \text{Binom}\bigl(1, \kappa(1 - \zeta)\bigr)}
For \eqn{\zeta = 0}, the thinning \eqn{\bullet} reduces to the binomial thinning
and for \eqn{\zeta = 1} to the Poisson thinning. The two thinnings of \eqn{E_t} are coupled via
\deqn{[\beta \circ E_t, (1 - \beta) \circ E_t] \sim \text{Mult}(E_t, \beta, 1 - \beta).}
The immigration process \eqn{I_t} consists of independently and identically distributed random variables which
can be Poisson, Hermite or negative binomial. This distribution is charaterized by its mean, denoted by
 \eqn{\tau}, and potentially a dispersion parameter \eqn{\psi}. See Bracher and Sobolova (2024) for details.
}
\details{
The function uses a variation of the forward algorithm to fit the models. This is
computationally expensive, meaning that the optimization can be quite lengthy
already for medium-level counts. If the time series contains values above ~15
moment-based estimation using `fit_inarma_moments` may be more practical.

Note that to improve convergence behaviour the function uses some slight regularization
which prevents \eqn{\beta > 0.97} and \eqn{\psi < 0.03}. If \eqn{\psi} is nonetheless
small, it may be more suitable to fit a Poisson model rather than a negative
binomial or Hermite.
}
\examples{
data("measles")
X <- measles$value
# Note: running the fit takes a little while.
\dontrun{
fit <- fit_inarma(X, family = "Poisson", offspring = "binomial")
summary(fit)
plot(fit, type = "fit")
}


}
