% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ingarch_type_models.R
\name{fit_ingarch}
\alias{fit_ingarch}
\title{Fitting an INGARCH model}
\usage{
fit_ingarch(
  observed,
  family = "Poisson",
  start = NULL,
  return_se = TRUE,
  control_optim = NULL
)
}
\arguments{
\item{observed}{a vector containing the observed count values}

\item{family}{the distributional family; one of `"Poisson"`, `"Hermite"` or `"NegBin"`}

\item{start}{initial values for the optimization routine (on the internal
scale; check the element `"optim"` of the return list).}

\item{return_se}{should standard errors be returned?}

\item{control_optim}{a list of options passed to `optim`}
}
\value{
a list with the following elements:
\describe{
\item{family}{the distribution family used.}
\item{offspring}{the offspring disribution used (always "Poisson" for INGARCH models).}
\item{coefficients_raw}{the estimated model coefficients on the internal scale.}
\item{se_raw}{if `return_se == TRUE`: the estimated standard errors on the internal scale.}
\item{cov_raw}{covariance matrix of the estimates on the internal scale.}
\item{coefficients}{the estimated model parameters transformed back to the natural scale.}
\item{se}{the estimated standard errors transformed back to the natural scale.}
\item{observed}{the vector of observed values provided by the user.}
\item{lik_distr}{matrix containing for each time point the conditional probabilities for all
values in the support (given the past). This contains notably all likelihood contributions and
can serve to compute fitted values.}
\item{fitted_values}{the fitted values as obtained from `lik_distr`}
\item{pearson_residuals}{the Pearson residuals}
\item{psi}{estimate and standard error of the parameter \eqn{\psi} from the
 GLM-formulation of the INARCH model. Included for making the transitions between the two formulations easier.}
\item{dim}{the number of fitted parameters}
\item{loglikelihood}{the log-likelihood of the fitted model}
\item{AIC}{the resulting AIC}
\item{convergence}{indicates whether optimization converged}
\item{nobs}{the number of observations}
\item{optim}{return object of the call to `optim`}
\item{fitting_method}{the method used to fit the model, here `"maximum_likelihood"`.}
}
}
\description{
Maximum likelihood inference for IGARCH(1, 1) model as described in Bracher
and Sobolova (2024):
}
\details{
\deqn{X_t = \theta * \bigl((1 - \beta) \circ E_t + I_t\bigr)}
\deqn{E_t = \beta \circ E_{t - 1} + \kappa \star X_{t - 1}}
where \eqn{\circ} denotes binomial thinning and \eqn{\star} denotes the
Poisson thinning. The type of thinning \eqn{*} depends on the distributional
family. See 'Details'.

The two thinnings of \eqn{E_t} are coupled via
\deqn{[\beta \circ E_t, (1 - \beta) \circ E_t] \sim \text{Mult}(E_t, \beta, 1 - \beta).}
All other thinnings are independent of each other.
The immigration process \eqn{I_t} consists of independently and identically distributed random variables which
can be Poisson, Hermite or negative binomial. This distribution is charaterized by its mean, denoted by
 \eqn{\tau}, and potentially a dispersion parameter \eqn{\psi}. See Bracher and Sobolova (2024) for details.

This display of the INGARCH model uses thinning
operators, which substantially differs from the common formulation of the
model in literature. For fitting an INGARCH model in its usual form
\deqn{X_t \sim \text{Pois}(\lambda_t),}
\deqn{\lambda_t = \nu + \alpha X_{t - 1} + \beta \lambda_{t - 1},}
we recommend using the `tscount` package (Liboschik T., Fokianos. K, Fried R.
(2017)).


The definition of thinning \eqn{*} depends on the distributional family.
For the Poisson INGARCH(1, 1) model, we have \eqn{\theta * 1 = 1}.

For the Hermite INGARCH(1, 1) model we have
\deqn{\theta * 1 = \begin{cases}
                   1 \quad \text{with probability } 2 - \theta,\\
                   2 \quad \text{with probability } \theta - 1.
                   \end{cases}}
When fitting the Hermite INGARCH(1, 1) model, `fit_ingarch` function operates
internally with the overdispersion parameter \eqn{\psi}, from which
\eqn{\theta} is retrieved using formula
\deqn{\theta = \frac{2}{2 - \psi}}

Finally for the Negative binomial INGARCH(1, 1) model, the thinning \eqn{ * }
is defined as
\deqn{\theta * 1 \sim \text{Log}\left(\frac{1}{1 + \psi}\right),}
where \eqn{\text{Log}(\pi)} denotes the logarithmic distribution with the
probability mass function.
\deqn{p(z) =  \frac{(1 - \pi)^z}{-z\log(\pi)}}
When fitting the Hermite INGARCH(1, 1) model, `fit_ingarch` function operates
internally with the overdispersion parameter \eqn{\psi}, from which
\eqn{\theta} is retrieved using formula
\deqn{\theta = \frac{\psi}{\log(1 + \psi)}.}
}
\examples{
data("measles")
X <- measles$value
\dontrun{
fit <- fit_ingarch(X, family = "Poisson")
}


}
