% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_inar_herm}
\alias{llik_inar_herm}
\title{Evaluate log-likelihood for a Hermite INAR(1) model}
\usage{
llik_inar_herm(vect, tau, kappa, psi)
}
\arguments{
\item{vect}{the vector of observed values}

\item{tau, kappa, beta, psi}{parameters of the Hermite INAR(1) model in the
thinning-based formulation}
}
\value{
log-likelihood
}
\description{
Evaluate log-likelihood for a Hermite INAR(1) model
}
