% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{llik_inar_pois}
\alias{llik_inar_pois}
\title{Evaluate log-likelihood of a Poisson INAR(1) model}
\usage{
llik_inar_pois(vect, tau, kappa, zeta)
}
\arguments{
\item{vect}{the vector of observed values}

\item{tau, kappa, zeta}{parameters of the Poisson INAR(1) model}
}
\value{
log-likelihood
}
\description{
Evaluate log-likelihood of a Poisson INAR(1) model
}
