% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_inarma}
\alias{sim_inarma}
\title{Simulate from an INARMA(1,1) model}
\usage{
sim_inarma(
  tau,
  psi = NULL,
  phi = NULL,
  beta = NULL,
  kappa,
  zeta = NULL,
  lgt,
  offspring = c("binomial", "binomial-Poisson"),
  family = c("Poisson", "Hermite", "NegBin"),
  E1 = NULL
)
}
\arguments{
\item{tau, psi, phi, beta, kappa, zeta}{the model parameters. Note that `phi` and `beta`
characterize the same underlying parameter, with `phi = 1 - beta`. Only one
of them should be specified, but we keep both to ensure compatibility for
different settings.}

\item{lgt}{the length of the simulated time series.}

\item{offspring}{type of the thinning \eqn{\kappa \bullet X_t}; one of `"binomial"`, or `"binomial-Poisson"`.}

\item{family}{distribution of the imports/innovations; one of `"Poisson"`, `"Hermite"` or `"NegBin"`.}
}
\value{
A named list with the following elements:
\describe{
\item{X}{The process \eqn{X_t}}
\item{E}{The hidden process \eqn{E_t}}
\item{I}{The innovation process \eqn{I_t}}
}
}
\description{
Simulation from an INARMA(1,1) model. For initialization an (approximation)
of the stationary distribution of \eqn{E[1]}.
}
