% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{tm_A_to_E}
\alias{tm_A_to_E}
\title{Get transition matrix from A[t] = (E[t] - L[t]) to E[t + 1]}
\usage{
tm_A_to_E(kappa, X_tminus1, support_A, support_E)
}
\arguments{
\item{kappa}{the parameter kappa (offspring mean)}

\item{X_tminus1}{vector of probabilities for X[t - 1]; also implies the support}

\item{support_A}{the chosen support for E}
}
\value{
a matrix containing transition probabilities (values of A in rows, of E in columns)
}
\description{
Get transition matrix from A[t] = (E[t] - L[t]) to E[t + 1]
}
